Split the package group tests by what they are about
`group-packages.test.ts` was 944 lines covering four separate concerns.
It is now a folder of four files, each named for the question it answers,
none over 400 lines:

- prices and quantities: what the group's own edit form saves and shows
  back, including what it does with a figure nobody could have meant
- what may join a package: the same invariant checked at every door it
  can be reached by — the group form, add-listings, the listings API, and
  the children sub-form
- a package that hides what is inside: nothing it holds may be reached or
  advertised on its own
- pulling apart a package people have bought: the refusals, and when they
  lift

The set-up they share moved to `helpers.ts` beside them. That file is not
a test file, so its lines are counted for coverage where they were not
before, which surfaced an untested guard: `sellPackageTicket` throws when
a booking does not succeed, and nothing had ever made it do so. It has a
test now — a helper that quietly handed back an empty token would let
every test around it check a sale that never happened.

The TODO also expected a `biome.json` override to delete; there is none.
The 944 lines were under the 1,000-line hard ceiling, which is why the
file grew unnoticed. The ~400-line aim is the one it was missing.
