+++ TODO.md
@@ -1681,28 +1681,6 @@ priced bundle reaches checkout today.
 
 ---
 
-## Split the package group test file
-
-*Origin: reviewer suggestion (Codex) on PR #1968.*
-
-`test/integration/server/group-packages.test.ts` is about 940 lines, well over
-the ~400 the repository aims for, and it is on the `biome.json` grandfathered
-list. PR #1968 removed two journeys from it but did not split it.
-
-It divides cleanly along the concerns already grouped inside it: what may and
-may not join a package, price and quantity parsing, per-day overrides, the
-share/QR affordances a hidden member must not show, and the sold-hidden-package
-guard on delete and un-package. Splitting it into a `group-packages/` folder
-along those lines also lets mutation runs target one narrow file per source
-rather than the whole suite.
-
-Starting point: the `describe`-less `test(...)` blocks in that file group
-naturally by the helpers they use (`expectPackageRejected`,
-`expectAddListingRejected`, `hiddenPackageWithBooking`); delete its
-`biome.json` override entry when the split lands.
-
----
-
 ## Give the stripe-mock install lock the same shape as every other file lock
 
 *Origin: noticed while unifying the locks behind `scripts/lock-file.ts`.*
